refactor(server): extract opencode server manager
