+++ packages/server/src/server/agent/providers/opencode-agent.full-access.test.ts
@@ -8,7 +8,8 @@ import { createOpencodeClient } from "@opencode-ai/sdk/v2/client";
 
 import { createTestLogger } from "../../../test-utils/test-logger.js";
 import type { AgentStreamEvent } from "../agent-sdk-types.js";
-import { OpenCodeAgentClient, OpenCodeServerManager } from "./opencode-agent.js";
+import { OpenCodeAgentClient } from "./opencode-agent.js";
+import type { OpenCodeServerManagerLike } from "./opencode/server-manager.js";
 
 interface MockOpenCodeClientOptions {
   agents?: unknown[];
@@ -23,13 +24,14 @@ function createEventStream(events: unknown[]): AsyncGenerator<never> {
   })();
 }
 
-function mockServerManager(): void {
-  vi.spyOn(OpenCodeServerManager, "getInstance").mockReturnValue({
+function createFakeServerManager(): OpenCodeServerManagerLike {
+  return {
+    ensureRunning: vi.fn().mockResolvedValue({ port: 1234, url: "http://127.0.0.1:1234" }),
     acquire: vi.fn().mockResolvedValue({
       server: { port: 1234, url: "http://127.0.0.1:1234" },
       release: vi.fn(),
     }),
-  } as never);
+  };
 }
 
 function mockOpenCodeClient(options: MockOpenCodeClientOptions = {}) {
@@ -121,7 +123,7 @@ describe("OpenCode full-access mode", () => {
   });
 
   test("includes virtual full-access mode with dynamic OpenCode agents", async () => {
-    mockServerManager();
+    const serverManager = createFakeServerManager();
     mockOpenCodeClient({
       agents: [
         { name: "build", mode: "primary", hidden: false, description: "Build agent" },
@@ -129,7 +131,9 @@ describe("OpenCode full-access mode", () => {
       ],
     });
 
-    const client = new OpenCodeAgentClient(createTestLogger());
+    const client = new OpenCodeAgentClient(createTestLogger(), undefined, undefined, {
+      serverManager,
+    });
     const modes = await client.listModes({ cwd: "/tmp/project", force: false });
 
     expect(modes.map((mode) => mode.id)).toEqual(["build", "plan", "full-access", "paseo-custom"]);
@@ -140,10 +144,12 @@ describe("OpenCode full-access mode", () => {
   });
 
   test("reports full-access but sends prompts through OpenCode build agent", async () => {
-    mockServerManager();
+    const serverManager = createFakeServerManager();
     const { promptAsync } = mockOpenCodeClient();
 
-    const client = new OpenCodeAgentClient(createTestLogger());
+    const client = new OpenCodeAgentClient(createTestLogger(), undefined, undefined, {
+      serverManager,
+    });
     const session = await client.createSession({
       provider: "opencode",
       cwd: "/tmp/project",
@@ -161,13 +167,15 @@ describe("OpenCode full-access mode", () => {
   });
 
   test("auto-approves tool permissions in full-access without surfacing them", async () => {
-    mockServerManager();
+    const serverManager = createFakeServerManager();
     const { permissionReply } = mockOpenCodeClient({
       events: [toolPermissionEvent(), idleEvent()],
     });
     const receivedEvents: AgentStreamEvent[] = [];
 
-    const client = new OpenCodeAgentClient(createTestLogger());
+    const client = new OpenCodeAgentClient(createTestLogger(), undefined, undefined, {
+      serverManager,
+    });
     const session = await client.createSession({
       provider: "opencode",
       cwd: "/tmp/project",
@@ -190,13 +198,15 @@ describe("OpenCode full-access mode", () => {
   });
 
   test("keeps questions separate from full-access tool auto-approval", async () => {
-    mockServerManager();
+    const serverManager = createFakeServerManager();
     const { permissionReply, questionReply } = mockOpenCodeClient({
       events: [questionEvent(), idleEvent()],
     });
     const receivedEvents: AgentStreamEvent[] = [];
 
-    const client = new OpenCodeAgentClient(createTestLogger());
+    const client = new OpenCodeAgentClient(createTestLogger(), undefined, undefined, {
+      serverManager,
+    });
     const session = await client.createSession({
       provider: "opencode",
       cwd: "/tmp/project",
